Practica_Algos/streaming/publisher/requirements.txt

twitchio==1.2.3
loguru==0.5.3
google-cloud-pubsub==2.1.0